refactor: change parameter of streak calculation function in profile service
The streak calculation method in the profile service has been refactored to directly accept array of contributions from the past 28 days rather than a GitHub username. The username-based fetching of profiles is now handled in the profile controller. This makes the profile service method more flexible and separates the responsibilities more clearly.